+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+ - Added `DrawCommandsIndicesBuffer` for multidraw elements indirect.
  - Added `ResidentTexture`, `TextureHandle` and `Texture::resident_if_supported()` for bindless textures.
  - Buffers no longer require their content to be `Send` or `'static` (except for types with the `Any` suffix).